pyproject.toml
[project]
name = "hireloom-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-pymupdf4llm>=0.4.1",
    "langgraph>=0.6.6",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
    "sentence-transformers>=3.0.1",
    "python-multipart>=0.0.20",
]
